iOS: move physical_device_mixin to ios/device
