+++ lib/calabash/ios/device/physical_device_mixin.rb
